Update WebGL tutorial files
