--- webgl-examples/tutorial/sample1/webgl.css
@@ -1,7 +1,0 @@
-canvas {
-  border: 2px solid black;
-  background-color: black;
-}
-video {
-  display: none;
-}
