+++ webgl-examples/tutorial/sample3/index.html
@@ -10,7 +10,7 @@
       crossorigin="anonymous"
       defer
     ></script>
-    <script src="webgl-demo.js" defer></script>
+    <script src="webgl-demo.js" type="module" defer></script>
   </head>
 
   <body>
